+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Documentation for rocJPEG is available at
 
 * Dependencies - Updates to core dependencies
 * LibVA Headers - Use public headers
+* mesa-amdgpu-va-drivers - RPM Package available on RPM from ROCm 6.2
 
 ### Fixes
 
@@ -32,4 +33,4 @@ Documentation for rocJPEG is available at
   * amdgpu-core - `1:6.1.60100-1741643`
 * libva-dev - `2.7.0-2` / `2.14.0-1`
 * mesa-amdgpu-va-drivers - `1:24.1.0`
-* rocJPEG Setup Script - `V1.1.0`
+* rocJPEG Setup Script - `V1.2.0`